refactor: 도메인내 클래스 패키지 분리에 따른 테스트 패키지 분리
