+++ src/test/java/blackjack/domain/deck/DeckTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.deck;
 
 import blackjack.domain.deck.Card;
 import blackjack.domain.deck.Deck;
