+++ src/test/java/blackjack/domain/deck/KindTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.deck;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
