+++ src/test/java/blackjack/domain/participant/HandsScoreTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import blackjack.domain.participant.HandsScore;
 import org.junit.jupiter.api.DisplayName;
