+++ src/test/java/blackjack/domain/participant/PlayerTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import blackjack.domain.deck.Card;
 import blackjack.domain.deck.Deck;
